WIP new voice_assistant entity draft
Initial definition of the new voice_assistant entity.
This is used to start a first proof of concept implementation and for
verifying the functionality.

Integration-API enhancement with streaming audio for voice-assistant
feature.
The voice stream is sent over the same WebSocket connection, but using
binary WebSocket messages with protocol buffers.
Protobuf allows future enhancements if other binary data needs to be
exchanged.